tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/140.html

<!DOCTYPE html>
<html><head>
  <title>scheduler: SVG script nested inline in external</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<div id="log"></div>
<script>
  var t = async_test();
</script>
<svg>
<script xlink:href="scripts/include-1.js">
  <script>
    t.step(function() {log("inline script #1")});
  </script>
</script>
</svg>
<script>
onload = t.step_func(function() {
  assert_array_equals(eventOrder, ["inline script #1", "external script #1"]);
  t.done();
});
</script>
